# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Example MCP REST API",
    "version": "0.4.1",
    "description": "Generated OpenAPI schema for rmcp-template's direct REST surface. TEMPLATE: rename Example identifiers and action schemas when adapting. Auth modes: loopback/trusted-gateway deployments may have no local auth; mounted bearer mode uses RTEMPLATE_MCP_TOKEN; OAuth mode uses bearer JWTs. REST actions require their action-specific scopes when auth is mounted."
  },
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Default local development server"
    }
  ],
  "tags": [
    {
      "name": "health",
      "description": "Unauthenticated runtime probes"
    },
    {
      "name": "capabilities",
      "description": "REST route inventory"
    },
    {
      "name": "direct-rest",
      "description": "Preferred typed REST routes"
    },
    {
      "name": "legacy-actions",
      "description": "Deprecated action-envelope compatibility"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Liveness probe",
        "operationId": "getHealth",
        "security": [],
        "responses": {
          "200": {
            "description": "Server is alive",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                },
                "examples": {
                  "ok": {
                    "value": {
                      "status": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "OpenAPI schema",
        "operationId": "getOpenApiSchema",
        "security": [],
        "responses": {
          "200": {
            "description": "Generated OpenAPI schema for the REST surface",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/status": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Local runtime status",
        "operationId": "getLocalStatus",
        "security": [],
        "responses": {
          "200": {
            "description": "Runtime status with secrets redacted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/capabilities": {
      "get": {
        "tags": [
          "capabilities"
        ],
        "summary": "Direct REST route inventory",
        "operationId": "getCapabilities",
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "Supported direct REST routes and metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/greet": {
      "post": {
        "tags": [
          "direct-rest"
        ],
        "summary": "Run greet",
        "description": "Direct REST route over the shared service layer.",
        "operationId": "postGreet",
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "greet result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreetResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GreetRequest"
              },
              "examples": {
                "greet": {
                  "value": {
                    "name": "Alice"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/echo": {
      "post": {
        "tags": [
          "direct-rest"
        ],
        "summary": "Run echo",
        "description": "Direct REST route over the shared service layer.",
        "operationId": "postEcho",
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "echo result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EchoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EchoRequest"
              },
              "examples": {
                "echo": {
                  "value": {
                    "message": "Hello!"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/status": {
      "get": {
        "tags": [
          "direct-rest"
        ],
        "summary": "Run status",
        "description": "Direct REST route over the shared service layer.",
        "operationId": "getStatus",
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "status result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/help": {
      "get": {
        "tags": [
          "direct-rest"
        ],
        "summary": "Run help",
        "description": "Direct REST route over the shared service layer.",
        "operationId": "getHelp",
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "help result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HelpResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/example": {
      "post": {
        "tags": [
          "legacy-actions"
        ],
        "summary": "Deprecated action-envelope dispatch",
        "description": "Compatibility shim for older clients. New application/platform servers should expose direct product REST routes and reserve action dispatch for MCP.",
        "operationId": "dispatchExampleActionDeprecated",
        "deprecated": true,
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              },
              "examples": {
                "greet": {
                  "value": {
                    "action": "greet",
                    "params": {
                      "name": "Alice"
                    }
                  }
                },
                "echo": {
                  "value": {
                    "action": "echo",
                    "params": {
                      "message": "Hello!"
                    }
                  }
                },
                "status": {
                  "value": {
                    "action": "status",
                    "params": {}
                  }
                },
                "help": {
                  "value": {
                    "action": "help",
                    "params": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Action result. Shape depends on the requested action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Static bearer token in bearer mode; OAuth mode also uses bearer JWTs. Loopback and trusted-gateway modes may not require local auth."
      }
    },
    "schemas": {
      "ActionName": {
        "type": "string",
        "enum": [
          "greet",
          "echo",
          "status",
          "help"
        ],
        "description": "REST-capable action names from crates/rtemplate-contracts/src/actions.rs."
      },
      "ActionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionName"
          },
          "params": {
            "type": "object",
            "description": "Action-specific parameters. greet.name is optional; echo.message is required.",
            "additionalProperties": true,
            "default": {}
          }
        }
      },
      "GreetRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "description": "Name to greet. Omit to greet the world."
          }
        }
      },
      "EchoRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1,
            "description": "Message to echo back. Must not be empty."
          }
        }
      },
      "ActionResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/GreetResponse"
          },
          {
            "$ref": "#/components/schemas/EchoResponse"
          },
          {
            "$ref": "#/components/schemas/StatusResponse"
          },
          {
            "$ref": "#/components/schemas/HelpResponse"
          },
          {
            "$ref": "#/components/schemas/RestTruncationResponse"
          }
        ]
      },
      "GreetResponse": {
        "type": "object",
        "required": [
          "greeting",
          "target"
        ],
        "properties": {
          "greeting": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "server": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "EchoResponse": {
        "type": "object",
        "required": [
          "echo"
        ],
        "properties": {
          "echo": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "StatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "server": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "transport": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "const": "ok"
          }
        },
        "additionalProperties": false
      },
      "CapabilitiesResponse": {
        "type": "object",
        "required": [
          "server",
          "version",
          "preferred_rest_style",
          "supported_routes",
          "routes"
        ],
        "properties": {
          "server": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "preferred_rest_style": {
            "type": "string",
            "const": "direct_routes"
          },
          "supported_routes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "routes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RestRoute"
            }
          }
        },
        "additionalProperties": false
      },
      "RestRoute": {
        "type": "object",
        "required": [
          "method",
          "path",
          "auth",
          "description"
        ],
        "properties": {
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "action": {
            "type": [
              "string",
              "null"
            ]
          },
          "auth": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "HelpResponse": {
        "type": "object",
        "required": [
          "actions",
          "mcp_only_actions",
          "usage",
          "examples"
        ],
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionName"
            }
          },
          "mcp_only_actions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "usage": {
            "type": "string"
          },
          "examples": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "RestTruncationResponse": {
        "type": "object",
        "required": [
          "truncated",
          "error",
          "max_response_bytes",
          "hint"
        ],
        "properties": {
          "truncated": {
            "type": "boolean",
            "const": true
          },
          "error": {
            "type": "string",
            "const": "response exceeded REST response size limit"
          },
          "max_response_bytes": {
            "type": "integer",
            "minimum": 1
          },
          "hint": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated request lacks the required scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "x-template": {
    "source": "scripts/check-openapi.py",
    "action_metadata": "crates/rtemplate-contracts/src/actions.rs",
    "preferred_rest_style": "direct_routes",
    "rest_actions": [
      "greet",
      "echo",
      "status",
      "help"
    ],
    "direct_rest_routes": {
      "greet": {
        "method": "POST",
        "path": "/v1/greet"
      },
      "echo": {
        "method": "POST",
        "path": "/v1/echo"
      },
      "status": {
        "method": "GET",
        "path": "/v1/status"
      },
      "help": {
        "method": "GET",
        "path": "/v1/help"
      }
    },
    "action_costs": {
      "greet": "cheap",
      "echo": "cheap",
      "status": "cheap",
      "elicit_name": "cheap",
      "scaffold_intent": "moderate",
      "help": "cheap"
    },
    "mcp_only_actions": [
      "elicit_name",
      "scaffold_intent"
    ]
  }
}